+++ changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Product List routing
 - Product View component
 - Product Service - getProduct
+- Cart Routing module
 
 ## [0.4.0] - 2021-10-03
 ### Removed